# forestclient/api_default.go
/*
 * Forest Client API
 *
 * This is a simple API to retrieve information about Forest Clients
 *
 * API version: 1.0.0
 * Contact: c.gustafson@cgi.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package forestclient

import (
	"encoding/json"
	"log"
	"net/http"
)

func GetForestClient(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)

	rec, err := json.Marshal(CreateRandomForestClient())

	if err != nil {
		log.Printf("error: %s", err)
	} else {
		w.Write(rec)
	}
}

func QueryForestClients(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)

	rec, err := json.Marshal(CreateRandomForestClientList())

	if err != nil {
		log.Printf("error: %s", err)
	} else {
		w.Write(rec)
	}

}