# forestclient/model_forest_client.go
/*
 * Forest Client API
 *
 * This is a simple API to retrieve information about Forest Clients
 *
 * API version: 1.0.0
 * Contact: c.gustafson@cgi.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package forestclient

type ForestClient struct {
	Id string `json:"id"`

	ClientName string `json:"clientName"`

	TypeCode string `json:"typeCode"`

	IsActive bool `json:"isActive"`
}